backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.6.1
python-dotenv==1.0.0
google-generativeai==0.3.0
httpx==0.25.2
